Opportunistically parse resolution/size of disconnected monitors
Disconnected xrandr monitors might still have a resolution set (this
happens when you disconnect but not disable a monitor). They might also
have a size set, but I haven't seen this happen.

This modifies the output parsing code to set those if found. There
already was a test case with such a situation, so we just add additional
asserts to actually ensure this data gets parsed.